+++ stock_available/models/res_config.py
