+++ stock_available/views/product_view.xml
