+++ stock_available/views/res_config_view.xml
