+++ stock_available_immediately/__init__.py
@@ -18,4 +18,4 @@
 #
 ##############################################################################
 
-from . import product
+from . import models
