+++ stock_available_immediately/tests/__init__.py
@@ -1,1 +1,3 @@
+# -*- encoding: utf-8 -*-
+
 from . import test_stock_available_immediately
